fix: update CDK versions for Java examples to latest (#217)
Updated all Java examples to the latest CDK version.
We switched to use a version range in `pom.xml`,
so we will now always use the latest versions as they are available.

Also fixed the expected CFN output for the ECS examples tests
(the updated CDK versions brought with them more taggable resources in the CFN spec).

Fixes #216