chore: housekeeping and cleanup - gitignore, package.json (#570)
A few minor housekeeping chores to make the repo slightly easier to work
with. First, standardizing on having the language-specific `.gitignore` files in
the top-level language directory, rather than (inconsistently) repeated in each
sample. There are a few exceptions where a local `.gitignore` is used, but it's
the exception now rather than the rule.

Also, the build scripts for csharp and go currently do an `npm install` command
as part of the script, but there was no `package.json` in those
directories. With older versions of Node (e.g., 12), this fails but doesn't exit
with a non-zero status; on newer versions, this causes the whole build to fail.

Lastly, a lot of the Typescript samples didn't list `aws-cdk` as either a
dependency or devDependency, relying on the globally-installed CDK version. Fine
on our build machines, but can cause problems in local development, especially if a
recent release changes the cloud assembly schema version.